+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Updated default vite version to 4
+* Added check for nested public/outDir
 
 ## v0.2.0 (2023-09-11)
 * Simplified config generation